+++ ganesh/src/ganesh/embed/database/maintenance/create.sql
@@ -6,24 +6,24 @@ create table dbver (
 
 --Tabelas do sistema
 create table empresa(
-	id_empresa int primary key,
+	id_empresa bigint primary key,
 	codigo varchar(25) not null unique,
 	nome varchar(56) not null,
 	cnpj varchar(14) not null
 );
 
 create table caminhao(
-	id_caminhao int primary key,
-	id_empresa int not null,
+	id_caminhao bigint primary key,
+	id_empresa bigint not null,
 	codigo varchar(25) not null unique,
 	placa varchar(10) not null,
-	tipo int not null,
+	tipo bigint not null,
 	
 	constraint fkCaminhaoEmpresa foreign key (id_empresa) references empresa(id_empresa)
 );
 
 create table pessoa(
-	id_pessoa int primary key,
+	id_pessoa bigint primary key,
 	codigo varchar(25) not null unique,
 	nome varchar(56) not null,
 	documento varchar(14) not null,
@@ -32,18 +32,18 @@ create table pessoa(
 );
 
 create table motorista(
-	id_motorista int primary key,
-	id_empresa int not null,
-	id_pessoa int not null,
+	id_motorista bigint primary key,
+	id_empresa bigint not null,
+	id_pessoa bigint not null,
 	
 	constraint fkMotoristaEmpresa foreign key (id_empresa) references empresa(id_empresa),
 	constraint fkMotoristaPessoa foreign key (id_pessoa) references pessoa(id_pessoa)
 );
 
 create table entrega(
-	id_entrega int primary key,
-	id_motorista int not null,
-	id_caminhao int not null,
+	id_entrega bigint primary key,
+	id_motorista bigint not null,
+	id_caminhao bigint not null,
 	dia date not null,
 	
 	constraint fkEntregaMotorista foreign key (id_motorista) references motorista(id_motorista),
@@ -51,9 +51,9 @@ create table entrega(
 );
 
 create table cliente(
-	id_cliente int primary key,
+	id_cliente bigint primary key,
 	codigo varchar(25) not null unique,
-	responsavel int not null,
+	responsavel bigint not null,
 	endereco varchar(84) not null,
 	contato varchar(56) not null,
 	
@@ -61,27 +61,27 @@ create table cliente(
 );
 
 create table nota(
-	id_nota int primary key,
+	id_nota bigint primary key,
 	codigo varchar(25) not null unique,
-	id_entrega int not null,
-	id_cliente int not null,
+	id_entrega bigint not null,
+	id_cliente bigint not null,
 	
 	constraint fkNotaCliente foreign key (id_cliente) references cliente(id_cliente),
 	constraint fkNotaEntrega foreign key (id_entrega) references entrega(id_entrega)
 );
 
 create table produto(
-	id_produto int primary key,
+	id_produto bigint primary key,
 	codigo varchar(25) not null unique,
 	nome varchar(56),
 	complemento varchar(56),
-	pesavel int not null default 0 check (pesavel in (0, 1))
+	pesavel bigint not null default 0 check (pesavel in (0, 1))
 );
 
 create table embalagem(
-	id_embalagem int primary key,
-	id_produto int not null,
-	qtd int not null,
+	id_embalagem bigint primary key,
+	id_produto bigint not null,
+	qtd bigint not null,
 	
 	constraint fkEmbalagemProduto foreign key (id_produto) references produto(id_produto)
 );
@@ -89,7 +89,7 @@ create table embalagem(
 create table produto_nota(
 	id_embalagem int,
 	id_nota int,
-	qtd int not null,
+	qtd bigint not null,
 	
 	constraint pkProdutoNota primary key (id_embalagem, id_nota),
 	constraint fkProdutoNotaEmbalagem foreign key (id_embalagem) references embalagem(id_embalagem),
@@ -98,7 +98,7 @@ create table produto_nota(
 
 --Controle de Acesso
 create table grupo(
-	id_grupo int primary key,
+	id_grupo bigint primary key,
 	codigo varchar(20) not null unique,
 	descricao varchar(84) not null
 );
@@ -112,8 +112,8 @@ create table grupo_role(
 );
 
 create table usuario(
-	id_usuario int primary key,
-	id_pessoa int not null,
+	id_usuario bigint primary key,
+	id_pessoa bigint not null,
 	username varchar(25) not null unique,
 	senha varchar(25) not null,
 	nome varchar(56) not null,
